Add first macro manifest file
Ideas is to have a manifest file, based on the one from the codec export, that lets you easily install macros and extensions with appstore like tools.
Unlike the codec bundle, the manifest shouldnt need to be zipped. Using external file url instead of having to bundle files makes it easier to update extensions.